requirements-windows.txt
# Windows 兼容依赖 - 排除 ROCm/NVIDIA 特定包，使用标准 PyTorch
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==2.4.1
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.3.2
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorthief==0.2.1
darkdetect==0.8.0
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.42.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
icontract==2.7.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
llama-index-core==0.14.7
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-instrumentation==0.4.2
llama-index-llms-llama-cpp==0.5.1
llama-index-vector-stores-chroma==0.5.3
llama-index-workflows==2.10.3
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==24.2
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.5.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySide6==6.10.0
PySide6-Fluent-Widgets==1.9.2
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
PySideSix-Frameless-Window==0.7.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
shiboken6==6.10.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch>=2.5.0
torchaudio>=2.5.0
torchvision>=0.20.0
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uvicorn==0.38.0
# uvloop 不支持 Windows，已排除
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.0
yarl==1.22.0
zipp==3.23.0
